Create README.md
A software program or tool titled “Hacker” (version 5.0). It appears that it is used to hack social media accounts such as Instagram, Facebook, and TikTok, through the options available on the screen.

The section shows details of the canvas such as the username on Telegram “alaui”, with a hint of the presence of passwords in the tool with an amount of “2 million”. After that, the user is directed to the social network to be engaged with (Instagram, Facebook or TikTok) and a large number to be checked (one or three accounts). Then you need usernames.